+++ mplgrid/mplgrid.py
@@ -9,8 +9,8 @@ __all__ = ['grid', 'grid_dimensions']
 def _grid_dimensions(ax_aspect=1, figheight=9, nrows=1, ncols=1,
                      grid_height=0.715, grid_width=0.95, space=0.05,
                      left=None, bottom=None,
-                     endnote_height=0.065, endnote_space=0.01,
-                     title_height=0.15, title_space=0.01,
+                     endnote_height=0, endnote_space=0.01,
+                     title_height=0, title_space=0.01,
                      ):
     """ A helper to calculate the grid dimensions.
 
@@ -39,18 +39,18 @@ def _grid_dimensions(ax_aspect=1, figheight=9, nrows=1, ncols=1,
         The location of the bottom endnote axes in fractions of the figure height.
         The default of None places the axes in the middle of the figure.
         If the endnote_height=0 then the grid is located at the bottom coordinate instead.
-    endnote_height: float, default 0.065
+    endnote_height: float, default 0
         The height of the endnote axes in fractions of the figure height.
-        The default is the endnote is 6.5% of the figure height.
-        If endnote_height=0, then the endnote axes is not plotted.
+        For, example 0.07 means the endnote axis is 7% of the figure height.
+        If endnote_height=0 (default), then the endnote axes is not plotted.
     endnote_space : float, default 0.01
         The space between the grid and endnote axis in fractions of the figure height.
         The default space is 1% of the figure height.
         If endnote_height=0, then the endnote_space is set to zero.
-    title_height : float, default 0.15
+    title_height : float, default 0
         The height of the title axis in fractions of the figure height.
-        The default is the title axis is 15% of the figure height.
-        If title_height=0, then the title axes is not plotted.
+        For, example 0.15 means the title axis is 15% of the figure height.
+        If title_height=0 (default), then the title axes is not plotted.
     title_space : float, default 0.01
         The space between the grid and title axis in fractions of the figure height.
         The default space is 1% of the figure height.
@@ -136,7 +136,7 @@ def _grid_dimensions(ax_aspect=1, figheight=9, nrows=1, ncols=1,
     return dimensions
 
 
-def _draw_grid(dimensions, left_pad=0, right_pad=0, axis=True, ax_key='grid'):
+def _draw_grid(dimensions, left_pad=0, right_pad=0, axis=True, grid_key='grid'):
     """ A helper to create a grid of axes in a specified location
 
     Parameters
@@ -151,7 +151,7 @@ def _draw_grid(dimensions, left_pad=0, right_pad=0, axis=True, ax_key='grid'):
         and the title/endnote. The padding units are fractions of the figure width.
     axis : bool, default True
         Whether the endnote and title axes are 'on'.
-    ax_key : str, default grid
+    grid_key : str, default grid
         The dictionary key for the main axes in the grid.
 
     Returns
@@ -183,7 +183,7 @@ def _draw_grid(dimensions, left_pad=0, right_pad=0, axis=True, ax_key='grid'):
     axs = np.squeeze(np.array(axs).reshape((dims['nrows'], dims['ncols'])))
     if axs.size == 1:
         axs = axs.item()
-    result_axes = {ax_key: axs}
+    result_axes = {grid_key: axs}
 
     title_left = dims['left'] + left_pad
     title_width = dims['grid_width'] - left_pad - right_pad
@@ -203,14 +203,16 @@ def _draw_grid(dimensions, left_pad=0, right_pad=0, axis=True, ax_key='grid'):
             ax_endnote.axis('off')
         result_axes['endnote'] = ax_endnote
 
-    return fig, result_axes
+    if dims['title_height'] == 0 and dims['endnote_height'] == 0:
+        return fig, result_axes[grid_key]  # no dictionary if just grid
+    return fig, result_axes  # else dictionary
 
 
 def grid(ax_aspect=1, figheight=9, nrows=1, ncols=1,
          grid_height=0.715, grid_width=0.95, space=0.05,
          left=None, bottom=None,
-         endnote_height=0.065, endnote_space=0.01,
-         title_height=0.15, title_space=0.01, axis=True, ax_key='grid'):
+         endnote_height=0, endnote_space=0.01,
+         title_height=0, title_space=0.01, axis=True, grid_key='grid'):
     """ Create a grid of axes in a specified location
 
     Parameters
@@ -238,25 +240,25 @@ def grid(ax_aspect=1, figheight=9, nrows=1, ncols=1,
         The location of the bottom endnote axes in fractions of the figure height.
         The default of None places the axes in the middle of the figure.
         If the endnote_height=0 then the grid is located at the bottom coordinate instead.
-    endnote_height: float, default 0.065
+    endnote_height: float, default 0
         The height of the endnote axes in fractions of the figure height.
-        The default is the endnote is 6.5% of the figure height.
-        If endnote_height=0, then the endnote axes is not plotted.
+        For, example 0.07 means the endnote axis is 7% of the figure height.
+        If endnote_height=0 (default), then the endnote axes is not plotted.
     endnote_space : float, default 0.01
         The space between the grid and endnote axis in fractions of the figure height.
         The default space is 1% of the figure height.
         If endnote_height=0, then the endnote_space is set to zero.
-    title_height : float, default 0.15
+    title_height : float, default 0
         The height of the title axis in fractions of the figure height.
-        The default is the title axis is 15% of the figure height.
-        If title_height=0, then the title axes is not plotted.
+        For, example 0.15 means the title axis is 15% of the figure height.
+        If title_height=0 (default), then the title axes is not plotted.
     title_space : float, default 0.01
         The space between the grid and title axis in fractions of the figure height.
         The default space is 1% of the figure height.
         If title_height=0, then the title_space is set to zero.
     axis : bool, default True
         Whether the endnote and title axes are 'on'.
-    ax_key : str, default grid
+    grid_key : str, default grid
         The dictionary key for the main axes in the grid.
 
     Returns
@@ -272,7 +274,7 @@ def grid(ax_aspect=1, figheight=9, nrows=1, ncols=1,
                                   endnote_height=endnote_height, endnote_space=endnote_space,
                                   title_height=title_height, title_space=title_space,
                                   )
-    fig, ax = _draw_grid(dimensions, axis=axis, ax_key=ax_key)
+    fig, ax = _draw_grid(dimensions, axis=axis, grid_key=grid_key)
     return fig, ax