+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 ----------------------
 
 Changed
-- default ax_key changed from ax to grid
+- renamed ax_key to grid_key and changed the default to 'grid'
+- changed the defaults so there is no title or endnote
+- changed the returned result so if there is no endnote or title \
+it returns the axes (e.g. numpy array of axes) rather than a dictionary.
 
 Added
 - added a test for the figsize